Add resulting call method
This method will take a result or object, if it is a failing result,
return early, or otherwise run the block. If the block is successful, it
will return a success result, otherwise it will return a failure result.

It will also ensure that it returns the klass of result that is passed
into it if it's not a default result object.