Add breaker and runner
This adds two more modules, Breaker and Runner which do very similar
things with an important difference.

Runner will run the method on every object from `.values` in result
passed in, and WILL call the block no matter what.

Breaker will run the method on every object from `.values` UNTIL there
is a result that returns false. If any of them return false, it will not
call the block.

Additionally, this adds a :wrapper option into these methods and the
Resulting.call method. I am not sure if this is ideal, but in the case
of rails, save needs to be surrounded by an
ActiveRecord::Base.transaction to have any real value. (And to not use
errors as flow control). For now this seems to work, where there is a
default wrapper which is just an empty block that yields.