Add wrap method to Result
Instead of needing to constantly check whether or not a parameter is a
result or value, this commit adds a wrap method which is then used in
the handler. If it implements resultable, then it will return the object
itself. Otherwise, it wraps it in a Resulting::Result with a success of
true.

It could be argued that success could start out as false and build
toward true, but it feels like in general you assume your flow is going
to go well and are looking for the point at which it fails.

Additionally, this automatically bails if it sees a failure, so it is
biased toward that approach.